+++ watchtower/__init__.py
@@ -2,5 +2,5 @@
 WatchTower - Podman Container Management Service
 """
 
-__version__ = "1.10.2"
+__version__ = "1.11.0"
 __author__ = "WatchTower Contributors"